+++ packages/component-library/src/components/form/mt-help-text/mt-helptext.stories.ts
@@ -5,7 +5,7 @@ import type { Meta } from "@storybook/vue3";
 type MtHelpTextMeta = Meta<typeof MtHelpText>;
 
 export default {
-  title: "Components/Form/mt-help-text",
+  title: "Components/mt-help-text",
   component: MtHelpText,
   args: {
     text: "Help text",