+++ packages/component-library/src/components/form/mt-switch/mt-switch.vue
@@ -46,7 +46,6 @@
         :inheritance="!isInheritanceField ? 'none' : isInherited ? 'linked' : 'unlinked'"
         :has-error="!!error"
         :required="required"
-        :disabled="disabled"
         @update:inheritance="
           if (isInherited) {
             $emit('inheritance-remove');